Misc Refactors
