+++ src/OwlApp/app/src/main/java/uk/co/dcurrey/owlapp/ui/player/NewPlayerActivity.java
@@ -1,4 +1,4 @@
-package uk.co.dcurrey.owlapp;
+package uk.co.dcurrey.owlapp.ui.player;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -8,6 +8,8 @@ import android.widget.Button;
 import android.widget.EditText;
 
 import androidx.appcompat.app.AppCompatActivity;
+
+import uk.co.dcurrey.owlapp.R;
 
 public class NewPlayerActivity extends AppCompatActivity
 {
