+++ src/OwlApp/app/src/main/java/uk/co/dcurrey/owlapp/ui/skill/SkillDetailsActivity.java
@@ -1,11 +1,13 @@
 
-package uk.co.dcurrey.owlapp;
+package uk.co.dcurrey.owlapp.ui.skill;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.content.Context;
 import android.os.Bundle;
 import android.widget.TextView;
+
+import uk.co.dcurrey.owlapp.R;
 
 public class SkillDetailsActivity extends AppCompatActivity
 {
