+++ src/OwlApp/app/src/main/res/layout/activity_player_details.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PlayerDetailsActivity">
+    tools:context=".ui.player.PlayerDetailsActivity">
 
     <TextView
         android:id="@+id/textView3"
