+++ src/OwlApp/app/src/main/res/layout/activity_skill_details.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SkillDetailsActivity">
+    tools:context=".ui.skill.SkillDetailsActivity">
 
     <TextView
         android:id="@+id/textView4"
