+++ src/OwlApp/app/src/main/res/navigation/mobile_navigation.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/nav_home"
-        android:name="uk.co.dcurrey.owlapp.ui.home.HomeFragment"
+        android:name="uk.co.dcurrey.owlapp.ui.character.HomeFragment"
         android:label="@string/menu_home"
         tools:layout="@layout/fragment_home">
 
@@ -17,7 +17,7 @@
     </fragment>
     <fragment
         android:id="@+id/nav_home_second"
-        android:name="uk.co.dcurrey.owlapp.ui.home.HomeSecondFragment"
+        android:name="uk.co.dcurrey.owlapp.ui.character.HomeSecondFragment"
         android:label="@string/home_second"
         tools:layout="@layout/fragment_home_second">
         <action
